+++ DrivingLicenseAppAPI/src/main/java/csapat/DrivingLicenseAppAPI/repository/InstructorRepository.java
@@ -12,20 +12,20 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface InstructorRepository extends JpaRepository<Instructors, Integer> {
+public interface InstructorRepository extends JpaRepository<Instructors, Long> {
 
     @Procedure(name = "getAllIntructor", procedureName = "getAllIntructor")
     List<Instructors> getAllInstructor();
 
     @Procedure(name = "getInstructor", procedureName = "getInstructor")
-    Optional<Instructors> getInstructor(@Param("idIN") Integer id);
+    Optional<Instructors> getInstructor(@Param("idIN") Long id);
 
     @Procedure(name = "deleteInstructor", procedureName = "deleteInstructor")
-    void deleteInstructor(@Param("idIN") Integer id);
+    void deleteInstructor(@Param("idIN") Long id);
 
     @Procedure(name = "getInstructorBySearch", procedureName = "getInstructorBySearch")
-    List<Integer> getInstructorBySearch(@Param("fuelTypeIdIN") Integer fuelTypeId, @Param("schoolIdIN") Integer schoolId,  @Param("categoryIdIN") Integer categoryId);
+    List<Long> getInstructorBySearch(@Param("fuelTypeIdIN") Long fuelTypeId, @Param("schoolIdIN") Long schoolId,  @Param("categoryIdIN") Long categoryId);
 
     @Query("select s from Instructors i JOIN i.students s where i.id = ?1 and s.isDeleted = false")
-    Page<Students> getAllStudents(Integer id, Pageable pageable);
+    Page<Students> getAllStudents(Long id, Pageable pageable);
 }
